#56 ログイン・新規登録のコンポーネントを作成
